Refactor: [BADA-200] 좋아요 기능 커스텀 훅으로 분리 및 로딩 상태 추가 #118
